requirements.txt
asgiref==3.6.0
Django==3.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
mongoengine==0.23.1
pydantic==1.10.2
pymongo==3.12.1
pytz==2022.6
sentencepiece==0.1.97
sqlparse==0.2.4
typing_extensions==4.4.0
tzdata==2022.7
waitress==2.1.2
wfastcgi==3.0.0
requests
apscheduler
ldap3==2.9.1
